remove deprecations and change structure
