+++ addon.xml
@@ -10,6 +10,8 @@
         <import addon="script.module.simplejson" version="2.0.10" />
         <import addon="script.module.pyxbmct" version="1.1.4"/>
     </requires>
+    <extension point="xbmc.python.library" library="helper.py" />
+    <extension point="xbmc.python.script" library="helper.py" />
     <extension point="xbmc.service" library="default.py" start="login"/>
     <extension point="xbmc.addon.metadata">
